refactor is done, rip build its gonna fail once again :/
