+++ app/main/ExtensionFactory.js
