+++ app/main/TaskManager.js
