+++ app/main/WindowFactory.js
